fix(models): picker cost estimate parses legacy cost strings, keeps sub-cent precision
Models whose costs live only in the legacy display strings ("3.43",
"$3.43/M") priced as free in the composer pickers; the estimate now falls
back to parsing the first number from those strings. The estimate also
returns fractional cents (float calc) instead of ceiling to whole cents, so
cheap models show e.g. CHF 0.002 rather than collapsing to 0.00/0.01; the
gating-side estimate_cost_cents still rounds up to whole cents.

+++ lib/magus/chat/model.ex
@@ -502,7 +502,7 @@ defmodule Magus.Chat.Model do
     # model pickers (workbench + SPA) so users can gauge how expensive a request
     # is. nil for image/video models.
     calculate :request_cost_cents,
-              :integer,
+              :float,
               Magus.Chat.Model.Calculations.RequestCostCents do
       public? true
       description "Approximate CHF cents for a reference request (composer model pickers)."
